tab: Make the loading state less intrusive
Remove the icon and fade the spinner in slowly when showing. Add a small
timeout before fading in so it's not visible at all when the tab opens
instantly.

Make sure not to stop the spinner too early so it doesn't disappear
during the transition.

Fixes https://gitlab.gnome.org/ZanderBrown/kgx/-/issues/54