pages: Add a context menu for tabs
Move the menu from menus.ui so that we can set it directly from the ui
file.

Make a new action group in the window with the prefix tab. Duplicate the
close action because we don't want to show a shortcut in the menu.

Fixes https://gitlab.gnome.org/ZanderBrown/kgx/-/issues/29