tab: use weak refs and destroy GTask
We were leaking multiple refs via GTask and unneeded strong refs

Additionally if the object were to die (or be removed from UI) we could
hit a race when :command is slow to spawn, or if it's death is out of
sync with ours. By using an explict weak pointer we can check if `self`
has been nulled out and thus abort.