application: don't match existing window size when restore is disabled
We'd fetch the to-restore window size from gsettings, which would be -1
x -1 if configured to not restore, but if there was already a window
open, we'd ignore that value and replace it with the active window size.
This somewhat defeats the purpose of the 'restore-window-size' setting.
Fix this by only trying to restore a window size, no matter from where,
if this setting is turned on (the default).