watcher: rework background state tracking
The ref counting system was error prone, and closing windows often
failed to clear their ref, leaving us permenantly in active polling.

Instead have KgxApplcation maintain a set of active windows, and ensure
they are cleared when leaving the applcation. This set in turn drives an
:in-background property, that is bound to an equivalent prop on watcher.

Whilst we are at it, drop some dead code.