ci: add a 32bit/clang build
Our flatpak builds both use GCC, and are on 64bit arches, so we don't
pick up on a range of possible issues until they hit distros.

To help limit this, do a 32bit build with clang.

Unfortunately dnf is apparently not very smart about i686 packages,
leading to a rather *looong* dependency list, but it gets the job done.

Whilst we are at it, produce a coverage report.