tab: initial title seperate from title
When creating a tab in an existing window, we copy over the title and
path of the current focused tab so as to avoid an annoying ‘flicker’
effect when the new tab near-instantly loads to probably the same (or
very similar), however this rather assumed a title would in fact be set,
and near-instantly, leading to some rather odd states when that didn't
actually happen.

We were also handling this initial title/path by simply setting the
‘real’ fields if they hadn't been already, which is a little nasty.

So, now we have duplicate properties, the ‘real’ values, and the initial
ones, leaving our consumers (KgxPages) to read the right ones, something
made quite easy with expressions. This is supported by the fact these
‘initial’ properties automatically clear themselves after a short
timeout, at which point if we never gained a ‘real’ value — sebeit, the
title is NULL, and again consumers can handle that as they see fit,
using a placeholder or otherwise.

Fix: https://gitlab.gnome.org/GNOME/console/-/issues/421