terminal: report errors as spads
Whilst journaling these errors was useful when developing KGX itself,
it's not so useful for end users to, say, click a link and get no
visible response.

Now when we encounter an error we raise a ‘spad’ indicating that
something didn't go to plan, hopefully avoiding leaving users in the
dark and possibly even in a good position to tell us what happened!